+++ app/javascript/packs/users/form.js
@@ -7,13 +7,18 @@ document.addEventListener("turbolinks:load", () => {
   const elements = document.querySelector(".new_user, .edit_user")
 
   if (elements) {
+    const name = document.querySelector("[v-model='user.name']").value
+    const language = document.querySelector("[v-model='user.language']:checked")
+    const other_language =
+      document.querySelector("[v-model='user.other_language']").value
+
     const app = new Vue({
       el: elements,
       data: {
         user: {
-          name: "",
-          language: undefined,
-          other_language: ""
+          name: name,
+          language: language ? language.value : undefined,
+          other_language: other_language
         }
       },
       mixins: [ TurbolinksAdapter ]